requirements.txt
torch==2.6.0
torchvision==0.21.0
efficientnet-pytorch==0.7.1
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
opencv-python==4.11.0.86
Pillow==11.1.0
streamlit==1.43.2
grad-cam==1.5.4
tqdm==4.67.1
scipy==1.15.2
